% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{ani_drawr}
\alias{ani_drawr}
\title{Animate drawing of a drawr dataframe}
\usage{
ani_drawr(dat)
}
\arguments{
\item{dat}{dataframe}

\item{x1}{column name of x1}

\item{y1}{column name of y1}

\item{x2}{column name of x2}

\item{y2}{column name of y2}

\item{join_x}{column name of join_x}

\item{join_y}{column name of join_y}
}
\value{
TRUE
}
\description{
Animate drawing of a drawr dataframe
}

