% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{define_orbit}
\alias{define_orbit}
\title{Define a planetary orbit}
\usage{
define_orbit(radius, offset, speed, parent_orbit = NULL)
}
\arguments{
\item{radius}{The radius of the orbit around its origin}

\item{offset}{The position of the orbit center relative to the absolute origin}

\item{speed}{The number of periods taken for a complete revolution}

\item{parent_orbit}{}
}
\description{
Define a planetary orbit
}

