% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{define_pantograph}
\alias{define_pantograph}
\title{Define a pantograph}
\usage{
define_pantograph(orbit1, orbit2, offset1 = 0, offset2 = 0,
  n_segments = 3, segment_length = 1)
}
\arguments{
\item{orbit1}{The first orbit it is attached to}

\item{orbit2}{The second orbit it is attached to}

\item{offset1}{The distance of the attachment from the origin of orbit1}

\item{offset2}{The distance of the attachment from the origin of orbit2}

\item{n_segments}{The number of segments in the pantograph}

\item{segment_length}{The length of each segment}
}
\value{
pantograph
}
\description{
Define a pantograph
}
\examples{
orbit1 <- define_orbit(c(0, 25), 1000)
orbit2 <- define_orbit(c(0, 8), 100, parent_orbit = orbit1)
orbit3 <- define_orbit(c(0, 1), 20, parent_orbit = orbit2)
pan <- define_pantograph(orbit3, orbit3, c(0, 1), c(0, 5), 3, 4)
}

