% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw_machine}
\alias{draw_machine}
\title{Draw basic state of drawing machine}
\usage{
draw_machine(pantograph)
}
\arguments{
\item{pantograph}{Object of class pantograph produced by `define_pantograph`}
}
\value{
ggplot objects
}
\description{
Draw basic state of drawing machine
}
\examples{
orbit1 <- define_orbit(c(0, 25), 1000)
orbit2 <- define_orbit(c(0, 8), 100, parent_orbit = orbit1)
orbit3 <- define_orbit(c(0, 1), 20, parent_orbit = orbit2)
pan <- define_pantograph(orbit3, orbit3, c(0, 1), c(0, 5), 3, 4)
draw_machine(pan)
}

