% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planetary.R
\name{eucl_dist}
\alias{eucl_dist}
\title{Get euclidean distance between two points}
\usage{
eucl_dist(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x position of point 1}

\item{y1}{y position of point 1}

\item{x2}{x position of point 2}

\item{y2}{y position of point 2}
}
\value{
numeric
}
\description{
Get euclidean distance between two points
}
\examples{
eucl_dist(0, 0, 1, 1)
}

