% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_complete_position}
\alias{get_complete_position}
\title{Get positions of complete orbit at theta = 0}
\usage{
get_complete_position(orbit, top_orbit = TRUE)
}
\arguments{
\item{orbit}{}
}
\value{

}
\description{
Get positions of complete orbit at theta = 0
}

