% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_drawing_point}
\alias{get_drawing_point}
\title{Get x of drawing point}
\usage{
get_drawing_point(x1, y1, x2, y2, segment_length, segment_number,
  return = "both")
}
\arguments{
\item{x1}{x position of point 1}

\item{y1}{y position of point 1}

\item{x2}{x position of point 2}

\item{y2}{y position of point 2}

\item{segment_length}{The length of scissor segments}

\item{segment_number}{The number of scissor segments}

\item{return}{a string specifying whether to return "x", "y" or "both"}
}
\value{
numeric
}
\description{
Get x of drawing point
}
\examples{
get_scissor_offset(0, 0, 1, 1, 2, 3)
}

