% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planetary.R
\name{get_hinge_y}
\alias{get_hinge_y}
\title{Get y of articulation point}
\usage{
get_hinge_y(x1, y1, x2, y2, dist, join_prop)
}
\arguments{
\item{x1}{x position of point1}

\item{y1}{y position of point1}

\item{x2}{x position of point2}

\item{y2}{y position of point2}

\item{dist}{Distance from each point to the articulation}

\item{join_prop}{Proportion of dist at which the arms join (between 0 and 1)}
}
\value{
numeric
}
\description{
Get y of articulation point
}

