% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_initial_theta}
\alias{get_initial_theta}
\title{Get the starting angle of an orbit relative to its parent}
\usage{
get_initial_theta(orbit)
}
\arguments{
\item{orbit}{Object of class orbit}
}
\value{
Angle in radians
}
\description{
Get the starting angle of an orbit relative to its parent
}
\examples{
orbit_base <- define_orbit(c(25, 0), 1000)
orbit2 <- define_orbit(c(8, 8), 100, parent_orbit = orbit_base)
get_initial_theta(orbit2)
}

