% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planetary.R
\name{get_join_x}
\alias{get_join_x}
\title{Get x of a hinged join between two points}
\usage{
get_join_x(x1, y1, x2, y2, dist)
}
\arguments{
\item{x1}{x position of point1}

\item{y1}{y position of point1}

\item{x2}{x position of point2}

\item{y2}{y position of point2}

\item{dist}{Distance from each point to the join}
}
\value{
numeric
}
\description{
Get x of a hinged join between two points
}

