% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_lims}
\alias{get_lims}
\title{Get axis limits to add to ggplot}
\usage{
get_lims(dat, var)
}
\arguments{
\item{dat}{A dataframe}

\item{var}{A variable name}
}
\value{
numeric vector of length 2
}
\description{
Get axis limits to add to ggplot
}

