% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_orbit_position}
\alias{get_orbit_position}
\title{Get (x, y) position of orbit location}
\usage{
get_orbit_position(theta, orbit)
}
\arguments{
\item{theta}{Angle in radians}

\item{orbit}{Object of class orbit}
}
\value{
numeric vector
}
\description{
Get (x, y) position of orbit location
}

