% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_scissor_offset}
\alias{get_scissor_offset}
\title{Get offset from the plane of two points for a given scissor set up}
\usage{
get_scissor_offset(x1, y1, x2, y2, segment_length, segment_number)
}
\arguments{
\item{x1}{x position of point 1}

\item{y1}{y position of point 1}

\item{x2}{x position of point 2}

\item{y2}{y position of point 2}

\item{segment_length}{The length of scissor segments}

\item{segment_number}{The number of scissor segments}
}
\value{
numeric
}
\description{
Get offset from the plane of two points for a given scissor set up
}
\examples{
get_scissor_offset(0, 0, 1, 1, 3, 1)
}

