% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planetary.R
\name{get_theta}
\alias{get_theta}
\title{Get theta of disc}
\usage{
get_theta(period, frequency)
}
\arguments{
\item{period}{Number of periods since beginning}

\item{frequency}{Number of periods taken for a complete revolution}
}
\value{
angle in radians
}
\description{
Get theta of disc
}
\examples{
get_theta(10, 100)
}

