% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_theta_diff}
\alias{get_theta_diff}
\title{Get angle of point 2 from point 1}
\usage{
get_theta_diff(pos1, pos2)
}
\arguments{
\item{pos1}{The first position, numeric length 2}

\item{pos2}{The second position, numeric length 2}
}
\value{
numeric
}
\description{
Get angle of point 2 from point 1
}
\examples{
get_theta_diff(c(0, 0), c(1, 1))
}

