% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_x}
\alias{get_x}
\title{Get x coordinates of orbit location}
\usage{
get_x(theta, rad, offset_x)
}
\arguments{
\item{theta}{Angle in radians}

\item{rad}{radius}

\item{offset_x}{x position of center}
}
\value{
numeric
}
\description{
Get x coordinates of orbit location
}
\examples{
get_x(pi / 6, 1, 0)
}

