% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planetary.R
\name{get_y}
\alias{get_y}
\title{Get y coordinates of planetary location}
\usage{
get_y(theta, rad, offset_y)
}
\arguments{
\item{theta}{Angle in radians}

\item{rad}{radius}

\item{offset_y}{y position of center}
}
\value{
numeric
}
\description{
Get y coordinates of planetary location
}
\examples{
get_y(pi / 6, 1, 0)
}

