% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggdrawr}
\alias{ggdrawr}
\title{Plot a drawr dataframe}
\usage{
ggdrawr(dat, join_x = "join_x", join_y = "join_y", print = TRUE)
}
\arguments{
\item{dat}{dataframe}

\item{join_x}{column name of join_x}

\item{join_y}{column name of join_y}
}
\value{
a ggplot2 object
}
\description{
Plot a drawr dataframe
}

