% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggdrawr_whole}
\alias{ggdrawr_whole}
\title{Plot a drawr dataframe}
\usage{
ggdrawr_whole(dat, x1 = "x1", y1 = "y1", x2 = "x2", y2 = "y2",
  join_x = "join_x", join_y = "join_y", print = TRUE)
}
\arguments{
\item{dat}{dataframe}

\item{x1}{column name of x1}

\item{y1}{column name of y1}

\item{x2}{column name of x2}

\item{y2}{column name of y2}

\item{join_x}{column name of join_x}

\item{join_y}{column name of join_y}
}
\value{
a ggplot2 object
}
\description{
Plot a drawr dataframe
}

