% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{orbit_transform}
\alias{orbit_transform}
\title{Transform a position by orbiting around an origin}
\usage{
orbit_transform(x1, y1, x2, y2, theta, return = "both")
}
\arguments{
\item{x1}{x position of point to be transformed}

\item{y1}{y position of point to be transformed}

\item{x2}{x position of origin}

\item{y2}{y position of origin}

\item{theta}{Angle in radians}

\item{return}{a string specifying whether to return "x", "y" or "both"}
}
\value{
numeric
}
\description{
Transform a position by orbiting around an origin
}
\examples{
orbit_transform(3,2,0,1,pi/4)
}

