% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{update_pantograph}
\alias{update_pantograph}
\title{Update a pantograph}
\usage{
update_pantograph(pantograph, orbit1 = NULL, orbit2 = NULL,
  offset1 = NULL, offset2 = NULL, n_segments = NULL,
  segment_length = NULL)
}
\arguments{
\item{pantograph}{Object of class pantograph, produced by `define_pantograph()`}

\item{orbit1}{Optional, object of class orbit}

\item{orbit2}{Optional, object of class orbit}

\item{offset1}{Optional, the distance of the attachment from the origin of orbit1}

\item{offset2}{Optional, the distance of the attachment from the origin of orbit2}

\item{n_segments}{Optional, the number of segments in the pantograph}

\item{segment_length}{Optional, the length of each segment}
}
\value{
pantograph
}
\description{
Update a pantograph
}

