% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{define_orbit}
\alias{define_orbit}
\title{Define a planetary orbit}
\usage{
define_orbit(offset, speed, parent_orbit = NULL)
}
\arguments{
\item{offset}{The position of the orbit center relative to the parent orbit center}

\item{speed}{The number of periods taken for a complete revolution}

\item{parent_orbit}{Name of an orbit object produced by `define_orbit()`}

\item{pantograph_point}{The position of the pantograph attachment relative to the orbit center}
}
\description{
Define a planetary orbit
}

