% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{eucl_dist}
\alias{eucl_dist}
\title{Get euclidean distance between two points}
\usage{
eucl_dist(pos1, pos2)
}
\arguments{
\item{pos1}{A numeric vector of length 2 defining the first position}

\item{pos2}{A numeric vector of length 2 defining the second position}
}
\value{
numeric
}
\description{
Get euclidean distance between two points
}
\examples{
eucl_dist(c(0, 0), c(1, 1))
}

