% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_drawing_point}
\alias{get_drawing_point}
\title{Get position of drawing point}
\usage{
get_drawing_point(pantograph, period_range = 1:100)
}
\arguments{
\item{pantograph}{An object of class pantograph}
}
\value{
A tibble
}
\description{
Get position of drawing point
}
\examples{
orbit_base <- define_orbit(c(0, 0), 1000)
orbit1 <- define_orbit(c(8, 0), 100, parent_orbit = orbit_base)
orbit2 <- define_orbit(c(1, 0), 20, parent_orbit = orbit_base)
orbit3 <- define_orbit(c(-8, 0), 90, orbit_base)
orbit4 <- define_orbit(c(-1, 0), 200, orbit3)
pan1 <- define_pantograph(orbit2, orbit4, c(0, 1), c(0, 5), 3, 4)
dp <- get_drawing_point(pan1)
}

