% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotters.R
\name{get_orbit_theta}
\alias{get_orbit_theta}
\title{Get theta of orbit}
\usage{
get_orbit_theta(period_range, orbit)
}
\arguments{
\item{period_range}{Integer vector of periods}

\item{orbit}{An object of class orbit}
}
\value{
angle in radians
}
\description{
Get theta of orbit
}
\examples{
get_theta(10, 100)
}

