% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planetary.R
\name{get_theta_diff}
\alias{get_theta_diff}
\title{Get angle of point 2 from point 1}
\usage{
get_theta_diff(x1, y1, x2, y2)
}
\arguments{
\item{x1}{x position of point 1}

\item{y1}{y position of point 1}

\item{x2}{x position of point 2}

\item{y2}{y position of point 2}
}
\value{
numeric
}
\description{
Get angle of point 2 from point 1
}
\examples{
get_theta_diff(0, 0, 1, 1)
}

